pyproject.toml
[tool.poetry]
name = "projeto-etl-spark"
version = "0.1.0"
description = "Projeto ETL com Spark"
authors = ["Your Name <your.email@example.com>"]
packages = [
    { include = "data" },
    { include = "main.py" }
]

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.0"
pyodbc = "^5.0.1"
pandas = "^2.1.4"
SQLAlchemy = "^2.0.23"
faker = "^22.6.0"
azure-storage-file-datalake = "^12.20.0"
pyspark = "~3.3.0"
delta-spark = "~2.3.0"
pyarrow = "^20.0.0"
azure-storage-blob = "^12.19.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"